+++ src/RadixDlt.NetworkGateway.GatewayApi/GatewayModelExtensions.cs
@@ -62,18 +62,42 @@
  * permissions under this License.
  */
 
-using System;
-using System.Threading.Tasks;
+using System.Diagnostics;
+using System.Linq;
 using CoreModel = RadixDlt.CoreApiSdk.Model;
 using GatewayModel = RadixDlt.NetworkGateway.GatewayApiSdk.Model;
 
-namespace RadixDlt.NetworkGateway.GatewayApi.Services;
+namespace RadixDlt.NetworkGateway.GatewayApi;
 
-public interface IPreviewServiceObserver
+internal static class GatewayModelExtensions
 {
-    ValueTask PreHandlePreviewRequest(GatewayModel.TransactionPreviewRequest request, string targetNode);
+    public static GatewayModel.TransactionCommittedOutcomeResponseFungibleChanges ToGatewayModel(this CoreModel.LtsEntityFungibleBalanceChanges input)
+    {
+        var feeBalanceChanges = input.FeeBalanceChanges
+            .Select(x => new GatewayModel.TransactionCommittedOutcomeResponseFungibleFeeBalanceChange(x.Type.ToGatewayModel(), x.ResourceAddress, x.BalanceChange))
+            .ToList();
 
-    ValueTask PostHandlePreviewRequest(GatewayModel.TransactionPreviewRequest request, string targetNode, GatewayModel.TransactionPreviewResponse response);
+        var nonFeeBalanceChanges = input.NonFeeBalanceChanges
+            .Select(x => new GatewayModel.TransactionCommittedOutcomeResponseFungibleBalanceChange(x.ResourceAddress, x.BalanceChange))
+            .ToList();
 
-    ValueTask HandlePreviewRequestFailed(GatewayModel.TransactionPreviewRequest request, string targetNode, Exception exception);
+        return new GatewayModel.TransactionCommittedOutcomeResponseFungibleChanges(input.EntityAddress, feeBalanceChanges, nonFeeBalanceChanges);
+    }
+
+    public static GatewayModel.TransactionCommittedOutcomeResponseNonFungibleChanges ToGatewayModel(this CoreModel.LtsEntityNonFungibleBalanceChanges input)
+    {
+        return new GatewayModel.TransactionCommittedOutcomeResponseNonFungibleChanges(input.EntityAddress, input.ResourceAddress, input.Added, input.Removed);
+    }
+
+    private static GatewayModel.LtsFeeFungibleResourceBalanceChangeType ToGatewayModel(this CoreModel.LtsFeeFungibleResourceBalanceChangeType input)
+    {
+        return input switch
+        {
+            CoreModel.LtsFeeFungibleResourceBalanceChangeType.FeePayment => GatewayModel.LtsFeeFungibleResourceBalanceChangeType.FeePayment,
+            CoreModel.LtsFeeFungibleResourceBalanceChangeType.FeeDistributed => GatewayModel.LtsFeeFungibleResourceBalanceChangeType.FeeDistributed,
+            CoreModel.LtsFeeFungibleResourceBalanceChangeType.TipDistributed => GatewayModel.LtsFeeFungibleResourceBalanceChangeType.TipDistributed,
+            CoreModel.LtsFeeFungibleResourceBalanceChangeType.RoyaltyDistributed => GatewayModel.LtsFeeFungibleResourceBalanceChangeType.RoyaltyDistributed,
+            _ => throw new UnreachableException($"Didn't expect {input} value"),
+        };
+    }
 }