+++ src/RadixDlt.NetworkGateway.GatewayApi/Handlers/DefaultTransactionHandler.cs
@@ -80,20 +80,23 @@ internal class DefaultTransactionHandler : ITransactionHandler
 {
     private readonly ILedgerStateQuerier _ledgerStateQuerier;
     private readonly ITransactionQuerier _transactionQuerier;
-    private readonly IPreviewService _previewService;
+    private readonly ITransactionPreviewService _transactionPreviewService;
+    private readonly ITransactionOutcomeService _transactionOutcomeService;
     private readonly ISubmissionService _submissionService;
     private readonly IOptionsSnapshot<EndpointOptions> _endpointConfiguration;
 
     public DefaultTransactionHandler(
         ILedgerStateQuerier ledgerStateQuerier,
         ITransactionQuerier transactionQuerier,
-        IPreviewService previewService,
+        ITransactionPreviewService transactionPreviewService,
+        ITransactionOutcomeService transactionOutcomeService,
         ISubmissionService submissionService,
         IOptionsSnapshot<EndpointOptions> endpointConfiguration)
     {
         _ledgerStateQuerier = ledgerStateQuerier;
         _transactionQuerier = transactionQuerier;
-        _previewService = previewService;
+        _transactionPreviewService = transactionPreviewService;
+        _transactionOutcomeService = transactionOutcomeService;
         _submissionService = submissionService;
         _endpointConfiguration = endpointConfiguration;
     }
@@ -133,7 +136,25 @@ internal class DefaultTransactionHandler : ITransactionHandler
 
     public async Task<GatewayModel.TransactionPreviewResponse> Preview(GatewayModel.TransactionPreviewRequest request, CancellationToken token = default)
     {
-        return await _previewService.HandlePreviewRequest(request, token);
+        return await _transactionPreviewService.HandlePreviewRequest(request, token);
+    }
+
+    public async Task<GatewayModel.TransactionCommittedOutcomeResponse> Outcome(GatewayModel.TransactionCommittedOutcomeRequest request, CancellationToken token = default)
+    {
+        var atLedgerState = await _ledgerStateQuerier.GetValidLedgerStateForReadRequest(request.AtLedgerState, token);
+        var committedTransaction = await _transactionQuerier.LookupCommittedTransaction(
+            request.IntentHash,
+            GatewayModel.TransactionDetailsOptIns.Default,
+            atLedgerState,
+            false,
+            token);
+
+        if (committedTransaction == null)
+        {
+            throw new TransactionNotFoundException(request.IntentHash);
+        }
+
+        return await _transactionOutcomeService.HandleOutcomeRequest(atLedgerState, committedTransaction, token);
     }
 
     public async Task<GatewayModel.TransactionSubmitResponse> Submit(GatewayModel.TransactionSubmitRequest request, CancellationToken token = default)